modelo_tour_faq:
1. Se arreglo el package de request.
