+++ src/main/java/com/tourya/api/controller/TourController.java
@@ -3,13 +3,10 @@ package com.tourya.api.controller;
 
 import com.tourya.api.common.PageResponse;
 import com.tourya.api.constans.enums.TourStatusEnum;
-import com.tourya.api.models.Tour;
-import com.tourya.api.models.responses.TourDetailsResponse;
 import com.tourya.api.models.responses.TourFullDataResponse;
 import com.tourya.api.models.responses.TourResponse;
-import com.tourya.api.models.resquest.TourCreateRequest;
-import com.tourya.api.models.resquest.TourFullDataRequest;
-import com.tourya.api.models.resquest.TourRequest;
+import com.tourya.api.models.request.TourFullDataRequest;
+import com.tourya.api.models.request.TourRequest;
 import com.tourya.api.services.TourService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
