+++ src/main/java/com/tourya/api/models/mapper/ProviderMapper.java
@@ -3,7 +3,7 @@ package com.tourya.api.models.mapper;
 
 import com.tourya.api.models.Provider;
 import com.tourya.api.models.responses.ProviderResponse;
-import com.tourya.api.models.resquest.RequestProviderRequest;
+import com.tourya.api.models.request.RequestProviderRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
