+++ src/main/java/com/tourya/api/models/mapper/TourMapper.java
@@ -7,9 +7,9 @@ import com.tourya.api.models.responses.TourFullDataResponse;
 import com.tourya.api.models.responses.TourIncludesExcludesResponse;
 import com.tourya.api.models.responses.TourMainAttractionResponse;
 import com.tourya.api.models.responses.TourResponse;
-import com.tourya.api.models.resquest.TourCreateRequest;
-import com.tourya.api.models.resquest.TourFullDataRequest;
-import com.tourya.api.models.resquest.TourRequest;
+import com.tourya.api.models.request.TourCreateRequest;
+import com.tourya.api.models.request.TourFullDataRequest;
+import com.tourya.api.models.request.TourRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
