+++ src/main/java/com/tourya/api/models/request/ProviderRequest.java
@@ -1,5 +1,4 @@
-package com.tourya.api.models.resquest;
-
+package com.tourya.api.models.request;
 
 import com.tourya.api.constans.enums.ProviderDocumentTypeEnum;
 import com.tourya.api.constans.enums.ProviderDocumentTypeEnumConverter;
@@ -9,27 +8,21 @@ import jakarta.persistence.Convert;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 
-@Getter
-@Setter
 @Data
-public class RequestProviderRequest {
+public class ProviderRequest {
     @NotEmpty(message = "name is mandatory")
     @NotNull(message = "name is mandatory")
     private String name;
 
-    @NotEmpty(message = "documentNumber is mandatory")
+    @NotEmpty(message = "documentNumber  is mandatory")
     @NotNull(message = "documentNumber is mandatory")
     private String documentNumber;
 
-    @NotNull(message = "documentType is mandatory")
     @Convert(converter = ProviderDocumentTypeEnumConverter.class)
     private ProviderDocumentTypeEnum documentType;
 
-    @NotNull(message = "serviceType is mandatory")
     @Convert(converter = ProviderServiceTypeEnumConverter.class)
     private ProviderServiceTypeEnum serviceType;
