+++ src/main/java/com/tourya/api/models/request/TourCategoryRequest.java
@@ -1,12 +1,12 @@
-package com.tourya.api.models.resquest;
+package com.tourya.api.models.request;
+
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 @Data
-public class TourRequest {
-
+public class TourCategoryRequest {
     @NotEmpty(message = "Name is mandatory")
     @NotNull(message = "Name is mandatory")
     private String name;
@@ -14,14 +14,4 @@ public class TourRequest {
     @NotEmpty(message = "Description is mandatory")
     @NotNull(message = "Description is mandatory")
     private String description;
-
-    @NotNull(message = "Tour Category ID is mandatory")
-    private Integer tourCategoryId;
-
-    private String duration;
-
-    private Integer maxPeople = 0; // Valor por defecto 0
-
-    private Integer highlight = 0; // Valor por defecto 0
-
 }
